test: split URL query param rule tests
